+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The **third number** is the patch version (bug fixes)
 
 - `harborapi.models.VulnerabilityItem.severity` field has had its type reverted to `Severity` from `Optional[Severity]`, which was a regression introduced in v0.16.0.
 - `harborapi.ext.regex.match` return type annotation is now correctly marked as `Optional[Match[str]]` instead of `Match[str]`.
+- `harborapi.HarborAsyncClient.get_retention_tasks` missing `limit` parameter in docstring.
 
 
 ## [0.16.1](https://github.com/pederhan/harborapi/tree/harborapi-v0.16.1) - 2023-04-24